Fix killing a killed ngrok process
If the ngrok process has been killed by another process, a call to
`ngrok.kill()` hangs indefinitely (never calling the callback).

Fix this by always binding the event handler to perform cleanup when
the child process exits. `ngrok.kill()` already does check if the
child process is `null` and calls the callback immediately, when that's
the case.